requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
packaging==23.2
PyMySQL==1.1.0
python-decouple==3.8
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
